+++ backend/src/main/java/com/amazon/repository/AccountRepository.java
@@ -4,9 +4,7 @@ import com.amazon.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
     Account findByEmail(String email);
 }
