+++ backend/src/main/java/com/amazon/repository/CustomerShoppingCartRepository.java
@@ -4,5 +4,5 @@ import com.amazon.entity.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
-    ShoppingCart findByAccount_Id(long accountId);
+    ShoppingCart findByAccount_Id(int accountId);
 }
